appium_tests/requirements.txt
Appium-Python-Client==3.1.1
selenium==4.18.1
pytest==8.1.1
pytest-html==4.1.1
